+++ pkg/dartdev/lib/src/commands/doc.dart
@@ -56,6 +56,9 @@ For additional documentation generation options, see the 'dartdoc_options.yaml' 
   }
 
   @override
+  String get category => 'Source code';
+
+  @override
   String get invocation => '${super.invocation} [<directory>]';
 
   @override